设置传感器模式，读取溶解氧数据

+++ demo_basic_0328/BSP/sensor/ph/ph.h
@@ -4,7 +4,7 @@
 __packed typedef struct  
 {										    
 		uint16_t version;					//版本数*100  放大了100倍,实际除以100
-		float PHvalue;					//PH值*100
+		uint16_t PHvalue;					//PH值*100
 		uint16_t voltage;					//电压值*100
 		uint16_t errUp;					//PH值报警上限*100
 		uint16_t errDown;					//PH值报警下限*100
